implemented the CTL model

+++ src/body.h
@@ -68,7 +68,7 @@ double fdRotKinEnergy(double,double,double,double);
 
 double fdDensityMassToRadius(double,double);
 double fdMassFromRadiusDensity(double,double);
-double fdMassToRad(double,double);
+double fdMassToRad(double,int);
 double fdRadToMass(double,double);
 double fdRadiusFreqToRotVel(double,double);
 double fdRadiusRotVelToFreq(double,double);
